Solar_Potential_Estimation/requirements.txt
# Core deep learning frameworks
torch>=1.11.0
torchvision>=0.12.0

# Data processing
numpy>=1.21.0
pillow>=8.0.0
opencv-python>=4.5.0
pandas>=1.3.0

# Image augmentation
albumentations>=1.1.0

# Machine learning utilities
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Model evaluation and optimization
pytorch-msssim>=0.2.1

# Configuration management
yacs>=0.1.8

# Visualization
matplotlib>=3.5.0

# HTTP requests
requests>=2.28.0

# Additional utilities
tqdm>=4.60.0

